requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-pinecone
langchain-huggingface
langchain-groq
pinecone
python-dotenv
pypdf
fastapi
sentence-transformers